+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- bug fix - specifying a component and its dependency as ids for bit remove was not working
+-
+
 ## [0.12.6] - 2018-02-27
 
 ### New